requirements.txt
flask==0.12.2
requests==2.18.0
urllib3==1.21.1
Werkzeug==0.11.10
PyYAML==3.12
Pillow==5.2.0
